backend/requirements.txt
Django==6.0.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
dj-database-url
gunicorn
whitenoise
Pillow
psycopg[binary]
